refactor: move `EVMUnionExec` to its own file
